FIX: TextInput - retaining cursor position when editing within word
