+++ code/ui.text/test/components/Test.tsx
@@ -1,28 +1,58 @@
 import * as React from 'react';
 import { Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
+import { filter, map, takeUntil } from 'rxjs/operators';
 
-import { css, t } from '../common';
-import { TestTextInput } from './Test.TextInput';
+import { Button, color, css, Hr, ITextInputProps, log, t, TextInput } from '../common';
+import { TestText } from './Test.Text';
 
-export type ITestProps = {};
+const LOREM =
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque nec quam lorem. Praesent fermentum, augue ut porta varius, eros nisl euismod ante, ac suscipit elit libero nec dolor. Morbi magna enim, molestie non arcu id, varius sollicitudin neque. In sed quam mauris. Aenean mi nisl, elementum non arcu quis, ultrices tincidunt augue. Vivamus fermentum iaculis tellus finibus porttitor. Nulla eu purus id dolor auctor suscipit. Integer lacinia sapien at ante tempus volutpat.';
 
-export class Test extends React.PureComponent<ITestProps, t.ITestState> {
-  public state: t.ITestState = {};
+export type ITestState = { value?: string };
+
+export class Test extends React.PureComponent<ITestState> {
+  public state: ITestState = {};
   private unmounted$ = new Subject();
-  private state$ = new Subject<Partial<t.ITestState>>();
+  private state$ = new Subject<Partial<ITestState>>();
+  private events$ = new Subject<t.TextInputEvent>();
+
+  private inputs: TextInput[] = [];
+  private inputRef = (ref: TextInput) => (this.inputs = ref ? [...this.inputs, ref] : this.inputs);
 
   /**
    * [Lifecycle]
    */
-
   public componentWillMount() {
+    const events$ = this.events$.pipe(takeUntil(this.unmounted$));
     const state$ = this.state$.pipe(takeUntil(this.unmounted$));
     state$.subscribe(e => this.setState(e));
+
+    events$.subscribe(e => {
+      log.info('🌳', e);
+    });
+
+    events$
+      .pipe(
+        filter(e => e.type === 'TEXT_INPUT/changing'),
+        map(e => e.payload as t.ITextInputChanging),
+      )
+      .subscribe(e => {
+        // e.cancel();
+      });
+
+    events$
+      .pipe(
+        filter(e => e.type === 'TEXT_INPUT/changed'),
+        map(e => e.payload as t.ITextInputChanged),
+      )
+      .subscribe(e => {
+        this.state$.next({ value: e.to });
+      });
   }
 
   public componentWillUnmount() {
     this.unmounted$.next();
+    this.unmounted$.complete();
   }
 
   /**
@@ -32,13 +62,79 @@ export class Test extends React.PureComponent<ITestProps, t.ITestState> {
     const styles = {
       base: css({
         Absolute: 0,
-        Flex: 'vertical',
+        display: 'flex',
+      }),
+      left: css({
+        width: 180,
+        backgroundColor: color.format(-0.04),
+        borderRight: `solid 1px ${color.format(-0.1)}`,
+        fontSize: 13,
+        padding: 10,
+        lineHeight: 1.8,
+      }),
+      right: css({
+        flex: 1,
+        position: 'relative',
       }),
     };
 
     return (
       <div {...styles.base}>
-        <TestTextInput />
+        <div {...styles.left}>
+          {this.button('value: <empty>', () => this.state$.next({ value: '' }))}
+          {this.button('value: short', () => this.state$.next({ value: 'hello' }))}
+          {this.button('value: long', () => this.state$.next({ value: LOREM }))}
+          <Hr margin={5} />
+          {this.button('focus', () => this.inputs[0].focus())}
+          {this.button('selectAll', () => this.inputs[0].selectAll().focus())}
+          {this.button('cursorToStart', () => this.inputs[0].cursorToStart().focus())}
+          {this.button('cursorToEnd', () => this.inputs[0].cursorToEnd().focus())}
+        </div>
+        <div {...styles.right}>{this.renderMain()}</div>
+      </div>
+    );
+  }
+
+  private button(title: string, handler?: () => void) {
+    return <Button label={title} onClick={handler} block={true} />;
+  }
+
+  private renderMain() {
+    const styles = {
+      base: css({
+        Absolute: 0,
+        padding: 30,
+        Scroll: true,
+      }),
+    };
+    return (
+      <div {...styles.base}>
+        <div>
+          {this.renderInput({ focusOnLoad: true })}
+          {this.renderInput({ maxLength: 5, placeholder: 'maxLength: 5' })}
+        </div>
+        <Hr margin={50} />
+        <TestText />
+      </div>
+    );
+  }
+
+  private renderInput(props: Partial<ITextInputProps>) {
+    const styles = {
+      base: css({
+        marginBottom: 20,
+      }),
+    };
+    return (
+      <div {...styles.base}>
+        <TextInput
+          ref={this.inputRef}
+          value={this.state.value}
+          placeholder={'TextInput'}
+          placeholderStyle={{ color: color.format(-0.2) }}
+          events$={this.events$}
+          {...props}
+        />
       </div>
     );
   }
